wgpu: Split off Frame into CommandRenderer and Frame - command renderer translates commands into draw calls on frame
